enrichment/requirements.txt
# enrichment/requirements.txt
functions-framework
google-cloud-storage
google-cloud-bigquery
google-genai
tenacity
pandas
pyarrow
db-dtypes
requests
pandas-ta